Plug-in Support Started
The absolute basics of plug-in support, by which I mean it can list
plug-ins, add plug-ins, and remove plug-ins. It cannot yet do anything
useful with them, partly because there are none to test it with, nor
is there any sort of UI of which to speak for the user to interact
with. This will all come in time. For now, we have a tiny "Languages"
controller for future basic testing.